pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-baton"
version = "0.1.0"
description = "Multi-agent orchestration system for Claude Code"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
baton = "agent_baton.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_baton*"]
